Rename modules_onehot and rewrite scripts
